# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InvoicePro API Documentation",
    "version": "1.0.0",
    "description": "Exhaustive OpenAPI 3.0 specification for all InvoicePro backend API endpoints (Authentication, Customer Management, Inventory & Stock adjustments, Invoice lifecycle workflows, and Admin User management)."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User Login",
        "description": "Logs a user in using credentials and returns a JWT access token and sets HttpOnly cookies.",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "example": "admin@enterprise.com"
                  },
                  "password": { "type": "string", "example": "admin123" }
                },
                "required": ["username", "password"]
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "admin@enterprise.com"
                  },
                  "password": { "type": "string", "example": "admin123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": { "type": "string" },
                    "refresh_token": { "type": "string" },
                    "token_type": { "type": "string", "example": "bearer" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "email": { "type": "string" },
                        "full_name": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get Current User Profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user profile information"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/customers": {
      "get": {
        "tags": ["Customers"],
        "summary": "List Customers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" },
            "description": "Search by customer name"
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          }
        }
      },
      "post": {
        "tags": ["Customers"],
        "summary": "Create Customer",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "Enterprise Corp" },
                  "email": {
                    "type": "string",
                    "example": "billing@enterprise.com"
                  },
                  "phone": { "type": "string" },
                  "billing_address": { "type": "string" },
                  "gst_number": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created successfully"
          }
        }
      }
    },
    "/api/customers/{id}": {
      "get": {
        "tags": ["Customers"],
        "summary": "Get Customer by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Customer details" },
          "404": { "description": "Customer not found" }
        }
      },
      "put": {
        "tags": ["Customers"],
        "summary": "Update Customer details",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string" },
                  "phone": { "type": "string" },
                  "billing_address": { "type": "string" },
                  "gst_number": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Customer updated" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}